+++ src/main/java/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/type/edge/EdgeType.java
@@ -10,10 +10,12 @@ import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
- * Model scheme type extension to register model as edge type (extends E). If model scheme is already create
+ * Model scheme type extension to register model as edge type (extends E). If model scheme is already created
  * in database, alters class to extend E (without data loss).
- * <p>Keeps track of object hierarchy: if annotated entity extend some other entity (e.g. VersionedEntity) then this
- * base class must extend edge type (E) and initializer will try to do it.</p>
+ * <p>Note: due to orient multiple inheritance support since 2.1, superclass is assigned directly
+ * to annotated class.</p>
+ * <p>If class or its super classes extend V error will be thrown. If class or any class in hierarchy
+ * already extend E, nothing will be done.</p>
  * <p>It's more safe to only annotate topmost classes.</p>
  * <p>Sample sql: alter class Model superclass E, create class Model extends E</p>
  *